Extract data column fetching into standalone functions
