+++ beacon-chain/p2p/custody.go
@@ -195,75 +195,6 @@ func (s *Service) AdmissiblePeersForDataColumns(
 	return dataColumnsByAdmissiblePeer, admissiblePeersByDataColumn, descriptions, nil
 }
 
-// SelectPeersToFetchDataColumnsFrom implements greedy algorithm in order to select peers to fetch data columns from.
-// https://en.wikipedia.org/wiki/Set_cover_problem#Greedy_algorithm
-func SelectPeersToFetchDataColumnsFrom(
-	neededDataColumns map[uint64]bool,
-	dataColumnsByPeer map[peer.ID]map[uint64]bool,
-) (map[peer.ID][]uint64, error) {
-	// Copy the provided needed data columns into a set that we will remove elements from.
-	remainingDataColumns := make(map[uint64]bool, len(neededDataColumns))
-	for dataColumn := range neededDataColumns {
-		remainingDataColumns[dataColumn] = true
-	}
-
-	dataColumnsFromSelectedPeers := make(map[peer.ID][]uint64)
-
-	// Filter `dataColumnsByPeer` to only contain needed data columns.
-	neededDataColumnsByPeer := make(map[peer.ID]map[uint64]bool, len(dataColumnsByPeer))
-	for pid, dataColumns := range dataColumnsByPeer {
-		for dataColumn := range dataColumns {
-			if remainingDataColumns[dataColumn] {
-				if _, ok := neededDataColumnsByPeer[pid]; !ok {
-					neededDataColumnsByPeer[pid] = make(map[uint64]bool, len(neededDataColumns))
-				}
-
-				neededDataColumnsByPeer[pid][dataColumn] = true
-			}
-		}
-	}
-
-	for len(remainingDataColumns) > 0 {
-		// Check if at least one peer remains. If not, it means that we don't have enough peers to fetch all needed data columns.
-		if len(neededDataColumnsByPeer) == 0 {
-			missingDataColumnsSortedSlice := uint64MapToSortedSlice(remainingDataColumns)
-			return dataColumnsFromSelectedPeers, errors.Errorf("no peer to fetch the following data columns: %v", missingDataColumnsSortedSlice)
-		}
-
-		// Select the peer that custody the most needed data columns (greedy selection).
-		var bestPeer peer.ID
-		for peer, dataColumns := range neededDataColumnsByPeer {
-			if len(dataColumns) > len(neededDataColumnsByPeer[bestPeer]) {
-				bestPeer = peer
-			}
-		}
-
-		dataColumnsSortedSlice := uint64MapToSortedSlice(neededDataColumnsByPeer[bestPeer])
-		dataColumnsFromSelectedPeers[bestPeer] = dataColumnsSortedSlice
-
-		// Remove the selected peer from the list of peers.
-		delete(neededDataColumnsByPeer, bestPeer)
-
-		// Remove the selected peer's data columns from the list of remaining data columns.
-		for _, dataColumn := range dataColumnsSortedSlice {
-			delete(remainingDataColumns, dataColumn)
-		}
-
-		// Remove the selected peer's data columns from the list of needed data columns by peer.
-		for _, dataColumn := range dataColumnsSortedSlice {
-			for peer, dataColumns := range neededDataColumnsByPeer {
-				delete(dataColumns, dataColumn)
-
-				if len(dataColumns) == 0 {
-					delete(neededDataColumnsByPeer, peer)
-				}
-			}
-		}
-	}
-
-	return dataColumnsFromSelectedPeers, nil
-}
-
 // custodyGroupsFromPeer compute all the custody groups indexed by peer.
 func (s *Service) custodyGroupsFromPeer(peers []peer.ID) (map[peer.ID]map[uint64]bool, error) {
 	peerCount := len(peers)
