+++ beacon-chain/sync/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "block_batcher.go",
         "broadcast_bls_changes.go",
         "context.go",
+        "data_columns.go",
         "data_columns_reconstruct.go",
         "data_columns_sampling.go",
         "deadlines.go",
@@ -131,6 +132,7 @@ go_library(
         "//time:go_default_library",
         "//time/slots:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
+        "@com_github_ethereum_go_ethereum//p2p/enode:go_default_library",
         "@com_github_hashicorp_golang_lru//:go_default_library",
         "@com_github_libp2p_go_libp2p//core:go_default_library",
         "@com_github_libp2p_go_libp2p//core/host:go_default_library",
@@ -163,6 +165,7 @@ go_test(
         "context_test.go",
         "data_columns_reconstruct_test.go",
         "data_columns_sampling_test.go",
+        "data_columns_test.go",
         "decode_pubsub_test.go",
         "error_test.go",
         "fork_watcher_test.go",
